Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -7,6 +7,6 @@ char* welcome() {
 int main() {
     float x;
     scanf("%f", &x);
-    printf("You entered: %lf",x);
+    printf("You entered: %lf",double(x));
     return 0;
 }
